♻️ Migrate the DTO model to the DTO data model module.
